Add a `mila-cpu` and `mila-gpu` entry in mila init
- Refactor some of the code of the `mila init` command.
- Add a "mila-cpu" and "mila-gpu" entry to ~/.ssh/config if the user
wants.
  - "mila-cpu": Asks for a cpu-only job in the unkillable partition,
with 2 cpus and 16 Gb of RAM.
  - "mila-gpu": Asks for a GPU job in the unkillable partition, with 2
cpus and 16 Gb of RAM.

The main goal of this is to avoid having users of the cluster click the
"mila" button in the VSCode Remote Explorer panel, as it connects them
to the login node. This way, they now directly have a way to connect to
a compute node.

This is fairly equivalent to using these parameters to `mila code`:
```
mila code . --alloc --partition=unkillable --cpus-per-task=2 --mem=16G
```

Signed-off-by: Fabrice Normandin <fabrice.normandin@gmail.com>