Fix bug and misleading type for `alloc` argument
`alloc` needs to be a list of strings, but it was typed as
`Sequence[str]`, which allows `str` to be passed (since `str`s are
sequences of `str`s).

This changes it to `list[str]` which is stricter and correct.

Incidentally, there was an undetected bug in the regression test at
`tests/integration/test_code_command.py::test_code` because I was
passing the allocation flags (a string) as salloc.

Signed-off-by: Fabrice Normandin <normandf@mila.quebec>